requirements.txt
aniso8601==9.0.1
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
bs4==0.0.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
email-validator==2.0.0.post2
Flask==2.3.2
flask-restx==1.1.0
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.2
numpy==1.24.3
osmium==3.6.0
pandas==2.0.1
pymongo==4.3.3
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.30.0
six==1.16.0
soupsieve==2.4.1
SQLAlchemy==2.0.15
typing_extensions==4.6.0
tzdata==2023.3
urllib3==2.0.2
Werkzeug==2.3.4
